Refactor the identity module some more
- Removes `ledger/identity.js`
- Creates `identity/index.js` as a convenient point to
re-export all the tiny pieces of identity
- Creates `identity/declaration.js` to hold the plugin declaration
- Creates `identity/v1.js` to hold the old (v1) identity type, and its
associated Graph manipulation methods
- Creates `identity/id.js` to export the IdentityId type
- Creates `identity/identityType.js` to export the identity type and
parser
- Updates clients to call `Identity` by `IdentityV1`

This micro-module approach is a good fit since we will have two versions
of the type which re-use most of the subtypes. Thanks to @Beanow for
demonstrating this pattern in #1508.

Part of #2109.

Test plan: `yarn test` passes.
