+++ src/ledger/identity/id.js
@@ -1,0 +1,5 @@
+// @flow
+
+import {type Uuid} from "../../util/uuid";
+
+export type IdentityId = Uuid;
