+++ src/ledger/identity/identityType.js
@@ -1,0 +1,11 @@
+// @flow
+
+import * as C from "../../util/combo";
+
+export type IdentityType = "USER" | "PROJECT" | "ORGANIZATION" | "BOT";
+export const parser: C.Parser<IdentityType> = C.exactly([
+  "USER",
+  "BOT",
+  "ORGANIZATION",
+  "PROJECT",
+]);
